Data.AffineSpace.Point

Description

A type for points (as distinct from vectors), with an appropriate AffineSpace instance.

Points

newtype Point v

Point is a newtype wrapper around vectors used to represent points, so we don't get them mixed up. The distinction between vectors and points is important: translations affect points, but leave vectors unchanged. Points are instances of the AffineSpace class from Data.AffineSpace.

Constructors

P v

unPoint :: Point v -> v

Convert a point p into the vector from the origin to p. This should be considered a "semantically unsafe" operation; think carefully about whether and why you need to use it. The recommended way to do this conversion would be to write (p .-. origin).

origin :: AdditiveGroup v => Point v

The origin of the vector space v.

(*.) :: VectorSpace v => Scalar v -> Point v -> Point v

Scale a point by a scalar.

mirror :: AdditiveGroup v => Point v -> Point v

Reflect a point through the origin.

Reflection through a point

relative :: AffineSpace p => p -> (Diff p -> Diff p) -> p -> p

Apply a transformation relative to the given point.

relative2 :: AffineSpace p => p -> (Diff p -> Diff p -> Diff p) -> p -> p -> p

Apply a transformation relative to the given point.

relative3 :: AffineSpace p => p -> (Diff p -> Diff p -> Diff p -> Diff p) -> p -> p -> p -> p

Apply a transformation relative to the given point.

reflectThrough :: AffineSpace p => p -> p -> p

Mirror a point through a given point.