Data.Functor.Context

Documentation

newtype Ctxt a

A value with a possible predecessor and successor.

This can be thought of as a limited version of a list zipper, showing only the direct neighbours.

Constructors

Ctxt

Fields

getCtxt :: (Maybe a, a, Maybe a)

toCtxt :: (Maybe a, a, Maybe a) -> Ctxt a

mapCtxt :: (a -> b) -> Ctxt a -> Ctxt b

extractCtxt :: Ctxt a -> a

addCtxt :: [a] -> [Ctxt a]

Add context to a sequence.

>>> addCtxt []
[]

>>> addCtxt [1..3]
[Ctxt {getCtxt = (Nothing,1,Just 2)},Ctxt {getCtxt = (Just 1,2,Just 3)},Ctxt {getCtxt = (Just 2,3,Nothing)}]