Data.Functor.Couple

Description

Defines two variants of (,) with lifted instances for the standard type classes.

The Functor, Applicative and Comonad instances are the standard instances. Applicative is used to lift Monoid and the standard numeric classes.

The only difference between Twain and Couple is the handling of Eq and Ord: Twain compares only the second value, while Couple compares both. Thus Couple needs an extra Ord b constraint for all sub-classes of Ord.

Documentation

newtype Twain b a

A variant of pair/writer with lifted instances for the numeric classes, using Applicative.

Constructors

Twain

Fields

getTwain :: (b, a)

newtype Couple b a

A variant of pair/writer with lifted instances for the numeric classes, using Applicative.

Constructors

Couple

Fields

getCouple :: (b, a)