Music.Parts

Description

Representation of musical instruments, parts and playing techniques.

Terminology

Parts represent a subset of a group of performers. It is mainly used for instrumental and vocal music, but some concetps may be useful in electronic music as well.

Section refers to a set of instrumentfamilies related by sound production method (i.e. woodwind).

Family refers to a set of instrument or voice types, which typically differ in size (i.e. saxophones).

Instrument refers to a set of instruments or voice types of a given type (i.e. soprano saxophones). Perhaps confusingly, this includes vocal types such as alto, tenor etc as well. However, there is no good general term that incorporate both instrument and voice type.

A Part is made up of an Instrument and a Division (i.e. Violin I). Solo parts are treated separately, so i.e. Violin solo II (as in a double concerto) is distinct from Violin II.

Subparts

module Music.Parts.Division

module Music.Parts.Subpart

Solo vs. tutti

module Music.Parts.Solo

Instruments

module Music.Parts.Instrument

module Music.Parts.Instrument.Brass

module Music.Parts.Instrument.Keyboard

module Music.Parts.Instrument.Percussion

module Music.Parts.Instrument.Strings

module Music.Parts.Instrument.Vocal

module Music.Parts.Instrument.Woodwind

Parts

module Music.Parts.Part

module Music.Parts.Group

Instruments etc

piccoloFlute :: Instrument

flute :: Instrument

altoFlute :: Instrument

bassFlute :: Instrument

oboe :: Instrument

corAnglais :: Instrument

heckelphone :: Instrument

ebClarinet :: Instrument

clarinet :: Instrument

aClarinet :: Instrument

bassClarinet :: Instrument

sopranoSax :: Instrument

altoSax :: Instrument

tenorSax :: Instrument

baritoneSax :: Instrument

bassoon :: Instrument

contraBassoon :: Instrument

horn :: Instrument

piccoloTrumpet :: Instrument

trumpet :: Instrument

cTrumpet :: Instrument

dTrumpet :: Instrument

bassTrumpet :: Instrument

altoTrombone :: Instrument

tenorTrombone :: Instrument

trombone :: Instrument

bassTrombone :: Instrument

tuba :: Instrument

timpani :: Instrument

snareDrum :: Part

kickDrum :: Part

crashCymbal :: Part

rideCymbal :: Part

clashCymbal :: Part

orchestralCymbal :: Part

piatti :: Part

piano :: Instrument

celesta :: Instrument

glockenspiel :: Instrument

vibraphone :: Instrument

marimba :: Instrument

xylophone :: Instrument

xylorimba :: Instrument

tubularBells :: Instrument

dulcimer :: Instrument

accordion :: Instrument

harmonica :: Instrument

violin :: Instrument

viola :: Instrument

cello :: Instrument

doubleBass :: Instrument

Parts

piccoloFlutes :: Part

flutes :: Part

altoFlutes :: Part

oboes :: Part

corAnglaises :: Part

clarinets :: Part

ebClarinets :: Part

bassClarinets :: Part

bassoons :: Part

contraBassoons :: Part

flutes1 :: Part

flutes2 :: Part

oboes1 :: Part

oboes2 :: Part

clarinets1 :: Part

clarinets2 :: Part

horns :: Part

highHorns :: [Part]

lowHorns :: [Part]

trumpets :: Part

trombones :: Part

trumpets1 :: Part

trumpets2 :: Part

trombones1 :: Part

trombones2 :: Part

tubas :: Part

violins :: Part

violins1 :: Part

violins2 :: Part

violas :: Part

cellos :: Part

doubleBasses :: Part

harp :: Part

claves :: Part

maracas :: Part

Ensembles (TODO move to separate module)

stringOrchestra :: [Part]