Music.Parts.Instrument

Description

A musical instrument.

You can think of this as a giant sum type:

data Instrument
 = Violin
 | Trumpet
 ...

All instruments in the MusicXML 3.1 Standard Sounds set are representable. For a full list, see:

See http://www.musicxml.com/for-developers/standard-sounds.

Partial mappings to MIDI instruments is available, see fromMidiProgram and toMidiProgram.

Documentation

data Instrument

An Instrument represents the set of all instruments of a given type.

Name

fullName :: Instrument -> Maybe String

Full instrument name.

shortName :: Instrument -> Maybe String

Instrument name abbrevation.

fromMidiProgram :: Int -> Instrument

Create an instrument from a MIDI program number. Given number should be in the range 0 - 127.

toMidiProgram :: Instrument -> Maybe Int

Convert an instrument to a MIDI program number. If the given instrument is not representable as a MIDI program, return Nothing.

fromMusicXmlSoundId :: String -> Instrument

Create an instrument from a MusicXML Standard Sound ID.

toMusicXmlSoundId :: Instrument -> Maybe String

Convert an instrument to a MusicXML Standard Sound ID. If the given instrument is not in the MusicXMl standard, return Nothing.

Clefs and transposition

transposition :: Instrument -> Interval

Transposition interval.

transpositionString :: Instrument -> String

A string representing transposition such as Bb or F.

standardClef :: Instrument -> Maybe Clef

Standard clef used for this instrument.

allowedClefs :: Instrument -> Set Clef

Clefs allowed for this instrument.

Playing range

playableRange :: Instrument -> Ambitus Interval Pitch

Playable range for this instrument.

comfortableRange :: Instrument -> Ambitus Interval Pitch

Comfortable range for this instrument.