Music.Parts.Part

Documentation

data Part

A part is a subdivided group of instruments of a given type.

_solo :: Lens' Part Solo

subpart :: HasSubpart a => Lens' a (SubpartOf a)

instrument :: Lens' Part Instrument

divide :: Int -> Part -> [Part]

Divide a part into n subparts.

containsPart :: Part -> Part -> Bool

a `containsPart` b holds if the set of players represented by a is an improper subset of the set of players represented by b.

smallestPart :: Part -> Part -> Part

smallestSubpart :: Subpart -> Subpart -> Subpart

largestPart :: Part -> Part -> Part

largestSubpart :: Subpart -> Subpart -> Subpart

distinctFrom :: Part -> Part -> Bool

Returns True iff x and y are completely distinct, i.e. neither contains the other.

>>> violins `distinctFrom` trumpets
True
>>> violins `distinctFrom` violins
False
>>> violins `distinctFrom` violins1
False
>>> violins1 `distinctFrom` violins
False
>>> violins1 `distinctFrom` violins2
True

allDistinct :: [Part] -> Bool

Returns True iff all given parts are distinct (as per distinctFrom).

solo :: Instrument -> Part

tutti :: Instrument -> Part