Music.Parts.Subpart

Description

A type to represent (recursive) subdivisions of a part.

Documentation

newtype Subpart

A subpart is a potentially infinite sequence of divisions, each typically designated using a new index type, i.e. I.1.2.

The empty subpart (also known as mempty) represents all the players of the group, or in the context of Part, all players of the given instrument.

Constructors

Subpart (NonEmpty Division)

containsSubpart :: Subpart -> Subpart -> Bool

properlyContainsSubpart :: Subpart -> Subpart -> Bool

isSubpartOf :: Subpart -> Subpart -> Bool

isProperSubpartOf :: Subpart -> Subpart -> Bool

class BoundIncr a where

Methods

maximum' :: NonEmpty a -> a

increment' :: a -> a

class BoundIncr (SubpartOf a) => HasSubpart a where

Associated Types

type SubpartOf a

Methods

subpart :: Lens' a (SubpartOf a)