Music.Pitch.Clef

Description

Clefs and staff positions.

Staff lines

data StaffLines

Represents staff number relative middle. Staff zero is either the middle staff, or if using an even number of lines, the staff below the middle space.

data HalfSpaces

Represents the difference betwee staff positions (often corresponding to one diatonic step).

type ClefLine = StaffLines

Clef representation

data ClefSymbol

Common clef symbols

Constructors

GClef

CClef

FClef

PercClef

NeutralClef

type ClefOctave = Integer

newtype Clef

Constructors

Clef

Fields

getClef :: (ClefSymbol, ClefOctave, ClefLine)

symbolName :: ClefSymbol -> String

Return the English name of the given clef.

symbolPitch :: ClefSymbol -> Maybe Pitch

Return the pitch implied by the given clef at the middle space or line.

positionPitch :: Clef -> StaffLines -> Maybe Pitch

Properties

isModernClef :: Clef -> Bool

Is this a clef used in contemporary notation?

isHistoricalClef :: Clef -> Bool

Is this an historical clef?

isVoiceClef :: Clef -> Bool

Is this a traditional voice clef, i.e. a C clef on some staff.

Standard clefs

trebleClef :: Clef

Standard treble clef.

bassClef :: Clef

Standard bass clef.

sopranoClef :: Clef

Standard soprano clef.

mezzoSopranoClef :: Clef

Standard mezzo soprano clef.

altoClef :: Clef

Standard alto clef.

tenorClef :: Clef

Standard tenor clef.

baritoneClef :: Clef

Standard baritone clef.

percussionClef :: Clef