Music.Pitch.Common.Harmony

Description

Basic harmony.

Documentation

isDissonance :: Interval -> Bool

Whether the given interval is a (harmonic) dissonance.

This does not check for augmented or dominished intervals, see isMelodicDissonance.

>>> isDissonance _M2
True

isConsonance :: Interval -> Bool

Whether the given interval is a (harmonic) consonance.

This does not check for augmented or dominished intervals, see isMelodicDissonance.

>>> isConsonance _P4
True

isPerfectConsonance :: Interval -> Bool

Whether the given interval is a perfect (harmonic) consonance.

This does not check for augmented or dominished intervals, see isMelodicDissonance.

isImperfectConsonance :: Interval -> Bool

Whether the given interval is an imperfect (harmonic) consonance.

This does not check for augmented or dominished intervals, see isMelodicDissonance.

isMelodicDissonance :: Interval -> Bool

Whether the given interval is a melodic dissonance.

A melodic dissonance is either augmented or diminished.

isMelodicConsonance :: Interval -> Bool

Whether an interval is melodic consonance.

A melodic consonance is either augmented or diminished.