Music.Pitch.Common.Interval

Description

Common intervals.

Intervals

data Interval

Interval type.

Creating intervals

interval :: Prism' (Quality, Number) Interval

View an interval as a pair of quality and number or vice versa.

>>> (Major, third)^?interval
Just _M3

>>> (Perfect, 4)^?interval
Just _P4

>>> (Perfect, 12)^?interval
Just _P12

This is a prism rather than an isomorphism, as only certain combinations of quality and number make sense.

>>> (Perfect, 3)^?interval
Nothing

>>> (Minor, 4)^?interval
Nothing

intervalTotalSteps :: Iso' (ChromaticSteps, DiatonicSteps) Interval

View an interval as a pair of total number of chromatic and diatonic steps.

>>> _P5^.from intervalTotalSteps
(7,4)

>>> d5^.from intervalTotalSteps
(6,4)

intervalAlterationSteps :: Iso' (ChromaticSteps, DiatonicSteps) Interval

View an interval as a pair of alteration and diatonic steps or vice versa.

>>> _P5^.from intervalAlterationSteps
(0,4)

>>> d5^.from intervalAlterationSteps
(-1,4)

steps :: Lens' Interval DiatonicSteps

View or set the number of chromatic steps in an interval.

alteration :: Lens' Interval ChromaticSteps

View or set the alteration (i.e. the number of chromatic steps differing from the excepted number) in an interval.

Synonyms

perfect :: Number -> Maybe Interval

Creates a perfect interval. If given an inperfect number, constructs a major interval.

major :: Number -> Maybe Interval

Creates a major interval. If given a perfect number, constructs a perfect interval.

minor :: Number -> Maybe Interval

Creates a minor interval. If given a perfect number, constructs a diminished interval.

augmented :: Number -> Maybe Interval

Creates an augmented interval.

diminished :: Number -> Maybe Interval

Creates a diminished interval.

doublyAugmented :: Number -> Maybe Interval

Creates a doubly augmented interval.

doublyDiminished :: Number -> Maybe Interval

Creates a doubly diminished interval.

Inspecting intervals

isNegative :: Interval -> Bool

Returns whether the given interval is negative.

>>> isNegative m3
False

>>> isNegative (-m3)
True

isPositive :: Interval -> Bool

Returns whether the given interval is positive.

isNonNegative :: Interval -> Bool

Returns whether the given interval is non-negative. This implies that it is either positive or a perfect unison.

isPerfectUnison :: Interval -> Bool

Returns whether the given interval a perfect unison.

>>> isPerfectUnison _P1
True

isStep :: Interval -> Bool

Returns whether the given interval is a step (a second or smaller).

Only diatonic number is taken into account, so _A2 is considered a step and m3 a leap, even though they have the same number of semitones.

>>> isLeap _A2
False

>>> isLeap m3
True

>>> isLeap m3
True

>>> isLeap _M2
False

>>> isLeap _P1
False

isLeap :: Interval -> Bool

Returns whether the given interval is a leap (larger than a second).

Only the diatonic number is taken into account, so _A2 is considered a step and m3 a leap, even though they have the same number of semitones.

>>> isLeap m3
True

>>> isLeap _M2
False

>>> isLeap _P1
False

Simple and compound intervals

isSimple :: Interval -> Bool

Returns whether the given interval is simple.

A simple interval is a non-negative interval spanning less than one octave.

>>> isSimple _P5
True

>>> filter isSimple [_P5, _M9, -m3, m3, _A7, _P8]
[_P5,m3,_A7]

isCompound :: Interval -> Bool

Returns whether the given interval is compound.

A compound interval is either a negative interval, or a positive interval spanning one octave or more. Note that compound intervals may be smaller than an octave if they are negative, so

>>> isCompound (-m3)
True
>>> isCompound m3
False

separate :: Interval -> (Octaves, Interval)

Separate a compound interval into octaves and a simple interval.

x*^_P8 + y = z  iff  (x, y) = separate z

>>> separate (2*^_P8+m3)
(2,m3)
>>> separate (3*^_P8+m3)
(3,m3)
>>> separate (0*^_P8+m3)
(0,m3)
>>> separate ((-1)*^_P8+m3)
(-1,m3)

simple :: Interval -> Interval

Returns the simple part of an interval.

(perfect octave)^*x + y = z  iff  y = simple z

class HasOctaves a where

Class of intervals that has a number of Octaves.

Methods

octaves :: a -> Octaves

Returns the number of octaves spanned by an interval.

The number of octaves is negative if and only if the interval is negative.

>>> octaves @Interval m3
0

>>> octaves @Interval _M9
1

>>> octaves @Interval (-m3)
-1

>>> octaves @Interval (d5 ^* 4)
2

octaves @Interval (-_P8)

1

Inversion

invert :: Interval -> Interval

Intervallic inversion.

The inversion an interval is determined as follows:

The number of a simple interval the difference of nine and the number of its inversion.

The quality of a simple interval is the inversion of the quality of its inversion.

The inversion of a compound interval is the inversion of its simple component.

>>> invert _P4
_P5

>>> invert _A4
d5

>>> invert m3
_M6

>>> invert (_P8 ^+^ m3)
_M6

>>> invert (-m3)
m3

Basis values

data IntervalBasis

This type represents standard basis for intervbals.

Constructors

Chromatic

Diatonic