Music.Pitch.Common.Number

Description

Number component of intervals.

Documentation

data Number

The number component of an interval (fourth, fifth) etc. May be negative, indicating a downward interval.

In keeping with music theory tradition, numbers are 1-indexed: the number 1 (or unison) represents an interval of 0 steps, 2 (or second) reprents an interval of 1 step, and so on.

class HasNumber a where

Methods

number :: a -> Number

Returns the number portion of an interval.

The interval number is negative if and only if the interval is negative.

See also quality, octaves and semitones.

>>> number (m3 :: Interval)
3

unison :: Number

A synonym for 1.

prime :: Number

A synonym for 1.

second :: Number

A synonym for 2.

third :: Number

A synonym for 3.

fourth :: Number

A synonym for 4.

fifth :: Number

A synonym for 5.

sixth :: Number

A synonym for 6.

seventh :: Number

A synonym for 7.

octave :: Number

A synonym for 8.

ninth :: Number

A synonym for 9.

tenth :: Number

A synonym for 10.

eleventh :: Number

A synonym for 11.

twelfth :: Number

A synonym for 12.

thirteenth :: Number

A synonym for 13.

fourteenth :: Number

A synonym for 14.

fifteenth :: Number

A synonym for 15.

diatonicSteps :: Iso' Number DiatonicSteps

>>> second^.diatonicSteps
1

>>> 2^.diatonicSteps
1

>>> 3^.from diatonicSteps
4