Music.Pitch.Common.Pitch

Description

Common pitch.

Accidentals

data Accidental

Accidental, represented as number of alterations. Generalized from natural and single/double sharp/flat to any number of steps.

natural :: Accidental

The natural accidental.

flat :: Accidental

The flat accidental.

sharp :: Accidental

The sharp accidental.

doubleFlat :: Accidental

The double flat accidental.

doubleSharp :: Accidental

The double sharp accidental.

Inspecting accidentals

isNatural :: Accidental -> Bool

Returns whether this is a natural accidental.

isFlattened :: Accidental -> Bool

Returns whether this is a flat, double flat etc.

isSharpened :: Accidental -> Bool

Returns whether this is a sharp, double sharp etc.

isStandardAccidental :: Accidental -> Bool

Returns whether this is a standard accidental, i.e. either a double flat, flat, natural, sharp or double sharp.

Name

data Name

Pitch name component.

Constructors

C

D

E

F

G

A

B

Pitch

data Pitch

Pitch type.

mkPitch :: Name -> Accidental -> Pitch

Creates a pitch from name accidental.

name :: Pitch -> Name

Returns the name of a pitch.

To convert a pitch to a numeric type, use octaves, steps or semitones on the relevant interval type, for example:

semitones ('a'' .-. c)

accidental :: Pitch -> Accidental

Returns the accidental of a pitch.

See also octaves, and steps and semitones.

>>> accidental cs
sharp

>>> accidental c
natural

>>> accidental (flatten cb)
doubleFlat

Diatonic and chromatic pitch

upDiatonicP :: Pitch -> DiatonicSteps -> Pitch -> Pitch

downDiatonicP :: Pitch -> DiatonicSteps -> Pitch -> Pitch

upChromaticP :: Pitch -> ChromaticSteps -> Pitch -> Pitch

downChromaticP :: Pitch -> ChromaticSteps -> Pitch -> Pitch

invertDiatonicallyP :: Pitch -> Pitch -> Pitch

invertChromaticallyP :: Pitch -> Pitch -> Pitch