Music.Pitch.Common.Quality

Description

Common interval quality.

Quality

data Quality

The quality component of an interval (minor, major, augmented). Generalized from single/double augmented/diminished to any number of steps.

Constructors

Augmented Integer

Major

Perfect

Minor

Diminished Integer

qualityTypes :: Quality -> [QualityType]

Return all possible quality types for a given quality.

>>> qualityTypes Perfect
[PerfectType]

>>> qualityTypes (Augmented 1)
[PerfectType,MajorMinorType]

isStandardQuality :: Quality -> Bool

Is this a quality used in standard notation

>>> filter isStandardQuality [Major,Minor,Perfect]
[Major,Minor,Perfect]

>>> filter isStandardQuality [Augmented n | n <- [1..4] ]
[Augmented 1,Augmented 2]

>>> filter isStandardQuality [Diminished n | n <- [1..4] ]
[Diminished 1,Diminished 2]

isSimpleQuality :: Quality -> Bool

Same as isStandardQuality but also disallow doubly augmented/diminished.

>>> filter isSimpleQuality [Augmented n | n <- [1..4] ]
[Augmented 1]

>>> filter isSimpleQuality [Diminished n | n <- [1..4] ]
[Diminished 1]

class HasQuality a where

Types of value that has an interval quality (mainly Interval and Quality itself).

Methods

quality :: a -> Quality

invertQuality :: Quality -> Quality

Invert a quality.

Perfect is unaffected, major becomes minor and vice versa, augmented becomes diminished and vice versa.

isPerfect :: HasQuality a => a -> Bool

Returns whether the given quality is perfect.

isMajor :: HasQuality a => a -> Bool

Returns whether the given quality is major.

isMinor :: HasQuality a => a -> Bool

Returns whether the given quality is minor.

isAugmented :: HasQuality a => a -> Bool

Returns whether the given quality is augmented (including double augmented etc).

isDiminished :: HasQuality a => a -> Bool

Returns whether the given quality is diminished (including double diminished etc).

Quality type

data QualityType

The actual alteration implied by a quality is dependent on whether it is attached to a major/minor vs. a perfect-style number. This type represents the two possibilities.

Constructors

PerfectType

MajorMinorType

expectedQualityType :: Number -> QualityType

The quality type expected for a given number, i.e. perfect for unisons, fourths, fifths and their compounds; major/minor for everything else.

isValidQualityNumber :: Quality -> Number -> Bool

Return whether the given combination of quality and number forms a valid interval expression.

Quality to alteration

data Direction

Constructors

Upward

Downward

qualityToAlteration :: Direction -> QualityType -> Quality -> Maybe ChromaticSteps

Return the alteration in implied by the given quality to a number of a given quality type.

>>> qualityToAlteration Upward MajorMinorType (Augmented 1)
Just 1

>>> qualityToAlteration Upward MajorMinorType Minor
Just -1

>>> qualityToAlteration Upward MajorMinorType (Diminished 1)
Just -2

>>> qualityToAlteration Upward PerfectType (Augmented 1)
Just 1

>>> qualityToAlteration Upward PerfectType Perfect
Just 0

>>> qualityToAlteration Upward PerfectType (Diminished 1)
Just -1

>>> qualityToAlteration Upward PerfectType Major
Nothing