Music.Pitch.Common.Spell

Description

Provides pitch spelling.

Spelling

About

Spelling type

type Spelling = Semitones -> Number

A spelling provide a way of notating a semitone interval such as tritone.

>>> spell usingSharps tritone
_A4

>>> spell usingFlats  tritone
d5

>>> spell modally     tone
_M2

spell :: HasSemitones a => Spelling -> a -> Interval

Spell an interval using the given Spelling.

spelled :: HasSemitones a => a -> Spelling -> Interval

Flipped version of spell. To be used infix, as in:

d5 `spelled` usingSharps

spellPitchRelative :: Tonic -> Spelling -> Pitch -> Pitch

Standard spellings

modally :: Spelling

Spell using the most the most common accidentals. Double sharps and flats are not preserved.

This spelling is particularly useful for modal music where the tonic is C.

c cs d eb e f fs g gs a bb b

usingSharps :: Spelling

Spell using sharps. Double sharps and flats are not preserved.

c cs d ds e f fs g gs a as b

usingFlats :: Spelling

Spell using flats. Double sharps and flats are not preserved.

c db d eb e f gb g ab a bb b

Simplifying pitches and intervals

useStandardQualities :: Interval -> Interval

Respell preserving general augmented/diminished diretion, but disallow all qualities except the standard ones.

Standard qualities include major, minor, perfect, augmented/diminished or doubly augmented/diminished.

useSimpleQualities :: Interval -> Interval

Same as useStandardQualities but disallow doubly augmented/diminished.

useStandardAlterations :: Tonic -> Pitch -> Pitch

Respell preserving general sharp/flat diretion, but disallow all qualities except the standard ones.

Standard qualities include natural, sharp, flat, double sharp and double flat.

useSimpleAlterations :: Tonic -> Pitch -> Pitch

Same as useStandardAlterations but disallow double sharp/flat.