Music.Pitch.Common.Types

Description

Common (Western classical) pitches, intervals and related types.

Even octaves and steps

data Octaves

Number of octaves. May be negative, indicating a downward interval.

data DiatonicSteps

Number of diatonic steps. May be negative, indicating a downward interval.

data ChromaticSteps

Number of chromatic steps. May be negative, indicating a downward interval.

type Semitones = ChromaticSteps

Number of semitones. May be negative, indicating a downward interval.

Intervals

data Number

The number component of an interval (fourth, fifth) etc. May be negative, indicating a downward interval.

In keeping with music theory tradition, numbers are 1-indexed: the number 1 (or unison) represents an interval of 0 steps, 2 (or second) reprents an interval of 1 step, and so on.

data Quality

The quality component of an interval (minor, major, augmented). Generalized from single/double augmented/diminished to any number of steps.

Constructors

Augmented Integer

Major

Perfect

Minor

Diminished Integer

data QualityType

The actual alteration implied by a quality is dependent on whether it is attached to a major/minor vs. a perfect-style number. This type represents the two possibilities.

Constructors

PerfectType

MajorMinorType

data IntervalBasis

This type represents standard basis for intervbals.

Constructors

Chromatic

Diatonic

data Interval

Interval type.

Pitch

data Name

Pitch name component.

Constructors

C

D

E

F

G

A

B

data Accidental

Accidental, represented as number of alterations. Generalized from natural and single/double sharp/flat to any number of steps.

data Pitch

Pitch type.