Music.Pitch.Scale

Description

Scales and chords.

Semantically is little distinction between a Scale and a Chord. Thus the Chord and Scale types are synonyms. We use a newtype wrapper to get a more idiomatic Inspectable instance.

Semantically Chord p and Scale p are countable subsets of some pitch space p.

A Mode (or 'ChordType) is like a Chord/Scale that has forgotten its origin.

TODO NEW

data Orientation

Constructors

Seq

Par

data Rooting

Constructors

NoRoot

Root

data ScaleChord :: Orientation -> Rooting -> Type -> Type -> Type where

Constructors

Mode :: NonEmpty v -> ScaleChord a 'NoRoot v p

ScaleChord :: p -> ScaleChord o 'NoRoot v p -> ScaleChord o 'Root v p

generator :: ScaleChord o r v p -> NonEmpty v

Returns the interval sequenc that generates the given mode.

The length of this sequence gives the number of notes in the scale, e.g. for a pentatonic (5-note) scale s, length (generator s) == 5, for a heptatonic (7-note) scale it is 7, and so on.

index :: AffinePair v p => ScaleChord o 'Root v p -> Integer -> p

member :: (Ord p, AffinePair v p) => ScaleChord o 'Root v p -> p -> Bool

tabulate :: AffinePair v p => ScaleChord o 'Root v p -> (Stream p, p, Stream p)

Modes and Chord types

type Mode = ScaleChord 'Seq 'NoRoot

type ChordType = ScaleChord 'Par 'NoRoot

leadingInterval :: AffinePair v p => Mode v p -> v

>>> leadingInterval majorScale
m2
>>> leadingInterval harmonicMinorScale
m2
>>> leadingInterval pureMinorScale
_M2

complementInterval :: AffinePair v p => ChordType v p -> v

>>> complementInterval majorTriad
_P4
>>> complementInterval minorTriad
_P4
>>> complementInterval majorMinorSeventhChord
_M2

invertMode :: AffinePair v p => Integer -> Mode v p -> Mode v p

length (generator x) == n -> invertMode n x = x

Scales

type Scale = ScaleChord 'Seq 'Root

scale :: AffinePair v p => p -> Mode v p -> Scale v p

scaleToList :: AffinePair v p => Scale v p -> [p]

Chords

type Chord = ScaleChord 'Par 'Root

chord :: AffinePair v p => p -> ChordType v p -> Chord v p

chordToList :: AffinePair v p => Chord v p -> [p]

Returns a single inversion of the given chord (no repeats!).

invertChord :: AffinePair v p => Integer -> ChordType v p -> ChordType v p

repeatingInterval :: AffinePair v p => Mode v p -> v

>>> repeatingInterval majorScale
_P8

Conversions

chordToScale :: Chord v p -> Scale v p

scaleToChord :: Scale v p -> Chord v p

chordTypeToMode :: ChordType v p -> Mode v p

modeToChordType :: Mode v p -> ChordType v p

reorient :: ScaleChord o r v p -> ScaleChord o' r v p

Common modes, scales and chords

Major-Minor/Common practice modes

majorScale :: Mode Interval Pitch

pureMinorScale :: Mode Interval Pitch

pureMinorScale = invertMode 5 majorScale
majorScale     = invertMode 2 pureMinorScale

harmonicMinorScale :: Mode Interval Pitch

melodicMinorScaleUp :: Mode Interval Pitch

Church/Gregorian modes

aeolian :: Mode Interval Pitch

locrian :: Mode Interval Pitch

ionian :: Mode Interval Pitch

dorian :: Mode Interval Pitch

phrygian :: Mode Interval Pitch

lydian :: Mode Interval Pitch

mixolydian :: Mode Interval Pitch

Other modes

majorPentaTonic :: Mode Interval Pitch

minorPentaTonic :: Mode Interval Pitch

bluesMinor :: Mode Interval Pitch

bluesMajor :: Mode Interval Pitch

bebopScale :: Mode Interval Pitch

The bebop scale, a.k.a. "bebop major".

wholeTone :: Mode Interval Pitch

octatonic :: Mode Interval Pitch

Modes of limited transposition

firstMode :: Mode Interval Pitch

secondMode :: Mode Interval Pitch

thirdMode :: Mode Interval Pitch

fourthMode :: Mode Interval Pitch

fifthMode :: Mode Interval Pitch

sixthMode :: Mode Interval Pitch

seventhMode :: Mode Interval Pitch

Triadic/Common practice chord types

majorTriad :: ChordType Interval Pitch

minorTriad :: ChordType Interval Pitch

augmentedChord :: ChordType Interval Pitch

diminishedChord :: ChordType Interval Pitch

halfDiminishedChord :: ChordType Interval Pitch

majorMinorSeventhChord :: ChordType Interval Pitch

Also known as "dominant seventh".

dominantSeventhChord :: ChordType Interval Pitch

majorMajorSeventhChord :: ChordType Interval Pitch

Also known as "major seventh".

minorMinorSeventhChord :: ChordType Interval Pitch

Also known as a "ii7". First inversion of major sixth.

minorMajorSeventhChord :: ChordType Interval Pitch

Also known as "major seventh".

frenchSixthChord :: ChordType Interval Pitch

germanSixthChord :: ChordType Interval Pitch

Non-triadic chord types

repeating :: v -> ChordType v p

Build a harmonic filed from repeating a single interval.

Up to enharmonic equivalence, we have:

repeating _m2 = chromaticCluster
repeating M2  = wholeToneCluster
repeating m3  = diminishedChord
repeating P4  = quartal

chromaticCluster :: ChordType Interval Pitch

wholeToneCluster :: ChordType Interval Pitch

quartal :: ChordType Interval Pitch

quintal :: ChordType Interval Pitch

Voiced chords/scales

data Voiced f v p

Constructors

Voiced

Fields

getChordScale :: f v p

getSteps :: NonEmpty Integer

getVoiced :: AffinePair v p => Voiced Chord v p -> NonEmpty p

Extract the pitches of a voiced chord.

voiced :: Chord v p -> Voiced Chord v p

The default closed voicing.

voiced (chord c majorTriad) = [c,e,g]

voiceIn :: Integer -> Chord v p -> Voiced Chord v p

A closed voicing with the given number of pitches, using the tonic as root.

voicedIn 5 (chord c majorTriad) = [c,e,g,c',e']

invertVoicing :: Integer -> Voiced Chord v p -> Voiced Chord v p

Invert a voiced chord.