Music.Score.Dynamics

Description

Provides functions for manipulating dynamics.

Manipulating dynamics

level :: Attenuable level dyn a => dyn -> a -> a

Set dynamic level, overwriting previous value.

louder :: Attenuable level dyn a => level -> a -> a

Increase dynamics (linear). For standard notation, switches dynamic marks the given number of step upwards.

>>> louder 1 (mp :: Dynamics)
_p

>>> louder 1 (ff :: Dynamics)
fff

>>> louder 5 [ppp,mf :: Dynamics]
[_f, fffff]

softer :: Attenuable v d a => v -> a -> a

Decrease dynamic (linear). For standard notation, switches dynamic marks the given number of step downwards.

cresc

Arguments

:: (HasPhrases' s a, Attenuable v d a, Fractional (Scalar v))

=> d

-> d
Initial dynamic.

-> s
Final dynamic.

-> s
Input value.

dim

Arguments

:: (HasPhrases' s a, Attenuable v d a, Fractional (Scalar v))

=> d

-> d
Initial dynamic.

-> s
Final dynamic.

-> s
Input value.

crescVoice :: (Attenuable l d a, Fractional (Scalar (GetLevel a))) => d -> d -> Voice a -> Voice a

volume :: (Attenuable v d a, Num d) => d -> a -> a

Scale dynamic values.

compressor

Arguments

:: (Attenuable l d a, Ord l, Num l)

=> d
Threshold

-> Scalar l
Ratio

-> a
Value to compress

-> a

Deprecated: Use compressUp (or compressDown)

compressUp

Arguments

:: (Attenuable v d a, Ord v, Num v)

=> d
Threshold

-> Scalar v
Ratio

-> a
Value to compress

-> a

Compress dynamics upwards.

>>> compressUp mp 2 [ppp,pp,_p,mp,mf,_f,ff,fff::Dynamics]
[ppp,pp,_p,mp,mp,fff,fffff,fffffff]

>>> compressUp 0 (1/2) (0.2 :: Amplitude)
Amplitude {getAmplitude = 0.1}

compressDown

Arguments

:: (Attenuable l d a, Ord l, Num l)

=> d
Threshold

-> Scalar l
Ratio

-> a
Value to compress

-> a

Compress dynamics downwards.

>>> compressUp mp 2 [ppp,pp,_p,mp,mf,_f,ff,fff::Dynamics]
[ppp,pp,_p,mp,mp,fff,fffff,fffffff]

>>> compressDown 0 1.5 (-0.2 :: Amplitude)
Amplitude {getAmplitude = -0.30000000000000004}

fadeIn :: (HasPosition a, HasDynamics' a, GetDynamic a ~ Behavior c, Fractional c) => Duration -> a -> a

Fade in.

fadeOut :: (HasPosition a, HasDynamics' a, GetDynamic a ~ Behavior c, Fractional c) => Duration -> a -> a

Fade in.

DynamicT

newtype DynamicT n a

Constructors

DynamicT

Fields

getDynamicT :: (n, a)

Attenuable class

type Attenuable level dyn a = (HasDynamics' a, level ~ GetLevel a, dyn ~ GetDynamic a, VectorSpace level, AffineSpace dyn)

Class of types that can be modified dynamically.

Dynamic type functions

type family GetDynamic (s :: Type) :: Type

Dynamics type.

type GetLevel a = Diff (GetDynamic a)

The GetLevel type represents difference in dynamics.

type family SetDynamic (b :: Type) (s :: Type) :: Type

Dynamic type.

type DynamicLensLaws' s t a b = (GetDynamic (SetDynamic a s) ~ a, SetDynamic (GetDynamic t) s ~ t, SetDynamic a (SetDynamic b s) ~ SetDynamic a s)

type DynamicLensLaws s t = DynamicLensLaws' s t (GetDynamic s) (GetDynamic t)

Accessing dynamics

class DynamicLensLaws s t => HasDynamics s t where

Class of types that provide a dynamic traversal.

Methods

dynamics :: Traversal s t (GetDynamic s) (GetDynamic t)

Access all dynamics.

class HasDynamics s t => HasDynamic s t where

Class of types that provide a single dynamic.

Methods

dynamic :: Lens s t (GetDynamic s) (GetDynamic t)

Access a single dynamic.

type HasDynamics' a = HasDynamics a a

type HasDynamic' a = HasDynamic a a

dynamic' :: (HasDynamic s t, s ~ t) => Lens' s (GetDynamic s)

Access a single dynamic.

dynamics' :: (HasDynamics s t, s ~ t) => Traversal' s (GetDynamic s)

Access all dynamics.

Context

vdynamic :: (HasDynamic' a, HasDynamic a b) => Lens (Voice a) (Voice b) (Voice (GetDynamic a)) (Voice (GetDynamic b))

View just the dynamices in a voice.

addDynCon :: (HasPhrases s t a b, HasDynamic a a, HasDynamic a b, GetDynamic a ~ d, GetDynamic b ~ Ctxt d) => s -> t