Music.Score.Export.StandardNotation

Description

This module defines a monomorphic representation of Western music notation.

Its main use to simplify the addition of notation-centric backends such as Sibelius, MusicXML, Lilypond, ABC notation and so on.

It generally follows the conventions of MusixXML but has a slightly more semantic flavor:

Bars, staves and rhytmical structure is explicit.

Spanners (legato, cresc dim etc) are represented by begin/end tags.

This is a consequence of representing the score as a tree rather than a graph, as spanners exist independently of the time hierarchy, and trees by definition do not allow multiple hierarchical relationships.

It could be interesting to use some kind of graph structure instead.

Harmoncics are represented as played, i.e. not by specifying note head shape etc explicitly.

Time and key signatures are global. Stave-specific key signatures, time signatures or tempi is not allowed.

The top level type is Work, which allow for representation of multi-movement concert pieces as well as film and theatre music.

Like Sibelius (and unlike MusicXML/Lilypond) we use a conceptual "system staff" that contains information pertaining to all parts, such as key and time signatures.

Documentation

data LabelTree b a

Constructors

Branch b [LabelTree b a]

Leaf a

foldLabelTree :: (a -> c) -> (b -> [c] -> c) -> LabelTree b a -> c

fromListLT :: Monoid b => [a] -> LabelTree b a

Common music notation

type BarNumber = Int

type TimeSignature = TimeSignature

type KeySignature = KeySignature

type RehearsalMark = RehearsalMark

type TempoMark = Tempo

data BracketType

Constructors

NoBracket

Bracket

Brace

Subbracket

type SpecialBarline = ()

data SystemBar

barNumbers :: Lens' SystemBar (Option (First BarNumber))

timeSignature :: Lens' SystemBar (Option (First TimeSignature))

keySignature :: Lens' SystemBar KeySignature

rehearsalMark :: Lens' SystemBar (Option (First RehearsalMark))

tempoMark :: Lens' SystemBar (Option (First TempoMark))

type SystemStaff = [SystemBar]

systemStaffLength :: SystemStaff -> Int

systemStaffTakeBars :: Int -> SystemStaff -> SystemStaff

type InstrumentShortName = String

type InstrumentFullName = String

type Transposition = Interval

type SibeliusFriendlyName = String

type SmallOrLarge = Any

type ScoreOrder = Sum Double

data StaffInfo

instrumentShortName :: Lens' StaffInfo InstrumentShortName

instrumentFullName :: Lens' StaffInfo InstrumentFullName

sibeliusFriendlyName :: Lens' StaffInfo SibeliusFriendlyName

instrumentDefaultClef :: Lens' StaffInfo Clef

transposition :: Lens' StaffInfo Transposition

smallOrLarge :: Lens' StaffInfo SmallOrLarge

scoreOrder :: Lens' StaffInfo ScoreOrder

data Pitch

Pitch type.

data ArpeggioNotation

Constructors

NoArpeggio
Don't show anything

NoArpeggioBracket
Show "no arpeggio" bracket

Arpeggio

UpArpeggio

DownArpeggio

data TremoloNotation

As written, i.e. 116-notes twice, can be represented as 18 note with 1 beams

These attach to chords

The distinction between MultiPitchTremolo and CrossBeamTremolo only really make sense for string and keyboard instruments.

Just n indicates n of extra beams/cross-beams. For unmeasured tremolo, use Nothing.

Constructors

MultiPitchTremolo (Maybe Int)

CrossBeamTremolo (Maybe Int)

NoTremolo

data BreathNotation

Constructors

NoBreath

Comma

Caesura

CaesuraWithFermata

type ArticulationNotation = ArticulationNotation

type DynamicNotation = DynamicNotation

type HarmonicNotation = (Any, Sum Int)

type SlideNotation = ((Any, Any), (Any, Any))

beginGliss :: Lens' SlideNotation Any

endGliss :: Lens' SlideNotation Any

beginSlide :: Lens' SlideNotation Any

endSlide :: Lens' SlideNotation Any

data Fermata

Constructors

NoFermata

Fermata

ShortFermata

LongFermata

type Ties = (Any, Any)

data Chord

A chord. A possibly empty list of pitches, played simultaneously. Naturally, an empty pitch list renders as a rest and a singleton list as a single note.

In MusicXML our chords are rendered using note elements. The MusicXML convention is to render each note or rest using a note. If the note is actually a rest, it is tagged with a rest/ element inside the note. If it is part of a chord, each note except one (typically the first) is tagged with a chord/ element.

It makes more sense if you think of the note element as representing a note/rest symbol and an implicit advancement of the time counter, and he chord element as an instruction to draw a note but skip advancing time (conversely backup/forward) do affect the time counter but do not draw anything.

TODO arguably ties, colors, harmonics and slide/gliss should be represented per notehead, and not per chord. This is simpler but may require more voices/layers to represent certain music.

TODO Better text. Separate different text types (again based on graphical presentation rather than semantics). Note that MusicXML unfortunately makes no distinction between expressive marks (i.e. dolce) versus instructions (i.e. pizz, sul pont).

pitches :: Lens' Chord [Pitch]

arpeggioNotation :: Lens' Chord ArpeggioNotation

tremoloNotation :: Lens' Chord TremoloNotation

breathNotation :: Lens' Chord BreathNotation

articulationNotation :: Lens' Chord ArticulationNotation

dynamicNotation :: Lens' Chord DynamicNotation

fermata :: Lens' Chord Fermata

chordColor :: Lens' Chord (Option (First (Colour Double)))

chordText :: Lens' Chord [String]

harmonicNotation :: Lens' Chord HarmonicNotation

slideNotation :: Lens' Chord SlideNotation

ties :: Lens' Chord Ties

newtype PitchLayer

A layer. A sequential composition of chords.

TODO we should replace Rhythm with something similar that also allows time-floating elements (i.e. grace notes).

Note that can not use Rhythm (Bool, Chord) (or similar) as that would scew up the time semantics of the Rhythm value, in other words: the floating nature of these notes should be captured in the time type.

Constructors

PitchLayer

Fields

getPitchLayer :: Rhythm Chord

data Bar

A bar. A parallel composition of layers.

All layers that make up a bar should have the same duration. An bar with no layers is allowed (this is mempty) and should be rendered as a bar rest.

MusicXML is a part/measure matrix (typically using column-major order) to allow it to be represented as a tree. Parts and measures are indexed by unique elements (called id and number respectively). Neither of these are rendered in the score (don't confuse with actual *visible* bar numbers), instead they are just token identifiers, and are conventionally named P1,P2,P3... and 1,2,3... respectively.

id="P1" number="1" ... /measure number="2" ... /measure /part ...

TODO free floating text/symbols?

Constructors

Bar

Fields

_clefChanges :: Map Duration Clef

_pitchLayers :: [PitchLayer]

clefChanges :: Lens' Bar (Map Duration Clef)

pitchLayers :: Lens' Bar [PitchLayer]

data Staff

A staff. A sequential composition of bars with meta-information.

By this we mean an actual staff on the page, whether it correlates with a single instrument or not (i.e. because it is part of a multi-staff part for i.e. piano, organ or percussion).

The hireachy we use here is clear: a score contains staves which are organized as a tree with bracket symbols at the branches. Traversing the tree gives us the staves top-to-botom.

MusicXML by contrast includes two separate notions of parts and staves.

The score is a part/measure matrix, and the part-list element provides a separate annotation for how parts are grouped (i.e. for brackets, barline convention and in-between staff names). In MusicXML staves are not part of the hierarchy, but are indicated separately: each part may specify a number of staves using

attribute staves N /staves /attribute

and certain elements (including notes note, forward and direction) [1] can be tagget to indicate which staff they should be placed on.

Consequently things like a piano staff can be represented in two ways: by using 2 (or more) explicist staves and a part-group, ur using a single part and the staff annotations. The single-part approach seem to be favored for piano music.

TODO we can also define multiple instrument parts using score-instrument, and then assign individual notes (and which other objects?) to this.

1

http://www.musicxml.com/tutorial/notation-basics/multi-part-music-2/)

Constructors

Staff

Fields

_staffInfo :: StaffInfo

_bars :: [Bar]

staffInfo :: Lens' Staff StaffInfo

bars :: Lens' Staff [Bar]

staffLength :: Staff -> Int

staffTakeBars :: Int -> Staff -> Staff

type Title = String

type Annotations = [(Span, String)]

type Attribution = Map String String

data MovementInfo

Constructors

MovementInfo

Fields

_movementTitle :: Title

_movementAnnotations :: Annotations

_movementAttribution :: Attribution

movementTitle :: Lens' MovementInfo Title

movementAnnotations :: Lens' MovementInfo Annotations

movementAttribution :: Lens' MovementInfo Attribution

data Movement

A movement. A parallel composition of staves grouped by various types of brackets, with meta-information.

Each staff (including the system-staff) should have the same number of bars (TODO backends should truncate to enforce this).

The system staff contains information that pertains to all simultaneous bars, such as key and time signature changes (implying that this model does not allow different time- or key singatures in different staves).

Constructors

Movement

Fields

_movementInfo :: MovementInfo

_systemStaff :: SystemStaff

_staves :: LabelTree BracketType Staff

movementInfo :: Lens' Movement MovementInfo

systemStaff :: Lens' Movement SystemStaff

staves :: Lens' Movement (LabelTree BracketType Staff)

movementAssureSameNumberOfBars :: Movement -> Movement

Pad all staves to the same number of bars.

All staves excepting the system staff must have a finite number of bars, or this will diverge. The system staff is not taken into account when calculating the number of bars (but it is padded if necessary).

data WorkInfo

title :: Lens' WorkInfo Title

annotations :: Lens' WorkInfo Annotations

attribution :: Lens' WorkInfo Attribution

data Work

A musical work. A sequential composition of movements with meta-information.

Called work to avoid confusion with Score.

Constructors

Work

Fields

_workInfo :: WorkInfo

_movements :: [Movement]

workInfo :: Lens' Work WorkInfo

movements :: Lens' Work [Movement]

MonadLog

class Monad m => MonadLog w m | m -> w where

A weaker form of MonadWriter which also supports imperative logging.

Minimal complete definition

Nothing

Methods

logger :: (a, w) -> m a

say :: w -> m ()

Pure export monad

data PureExportM a

Basic monad for exporting music.

Run as a pure computation with internal logging.

runPureExportM :: PureExportM a -> Either String (a, [String])

Run the computation.

Return the result and all emitted log messages, or an error message.

runPureExportMNoLog :: PureExportM a -> Either String a

Run the computation.

Return the result or an error message. The log messages are discarded.

IO export monad

data IOExportM a

Basic failure and logging in the IO monad.

Logs to standard output.

runIOExportM :: IOExportM a -> IO a

Standard notation

type Asp1 = Maybe Asp1a

type Asp1a = PartT Part (StaffNumberT (TremoloT (ColorT (TextT (HarmonicT (SlideT (TechniqueT SomeTechnique (ArticulationT Articulation (DynamicT Dynamics Pitch)))))))))

type Asp = Score Asp1

type StandardNotationExportM m = (MonadLog String m, MonadError String m)

toStandardNotation :: StandardNotationExportM m => Asp -> m Work

Lilypond backend

data LilypondLayout

Constructors

LilypondInline
Render all music on a single page, without titles (e.g. for inline examples in the documentation).

LilypondScore
Render normally.

LilypondBigScore
Render all music on a single page, with titles.

data LilypondOptions

Constructors

LilypondOptions

Fields

layout :: LilypondLayout

defaultLilypondLayout :: LilypondLayout

defaultLilypondOptions :: LilypondOptions

type LilypondExportM m = (MonadLog String m, MonadError String m)

toLy :: LilypondExportM m => LilypondOptions -> Work -> m (String, Music)

MusicXML backend

type MusicXmlExportM m = (MonadLog String m, MonadError String m)

toXml :: MusicXmlExportM m => Work -> m Score

MIDI backend

type MidiExportM m = (MonadLog String m, MonadError String m)

toMidi :: MidiExportM m => Asp -> m Midi

Convert the given music to a Standard MIDI File.

A MIDI file is a number of tracks to be composed in parallel.

Each track consists of timestamped messages. Most messages are one of the "normal" ones (on, off, pressure, control change, program change, pitch wheel), each assigned to a channel which is a number in [0..15]. While MIDI files are unlimited in the number of tracks, we are limited to 16 channels because that's how MIDI messages work (MIDI messages predate Standard MIDI Files).

We do not emit any SysEx messages.

We do not render many meta messages, only the ones required for correct play. E.g. no key signature changes etc (instead we transform the delta times before rendering to MIDI). The MIDI backend is meant to be used for sound rendering only, *not* as a logical/score-level export format.

For now we do not support more than 16 sounds types. We only support one sound per instrument (the "standard" one, e.g. arco for violin). We perform no reallocation of channels, instead we check how many instruments there are in the given score and fail if there are more than 15 (we follow convention in reserving channel 10 for percussion, which we currently don't render). We then allocate a channel for each sound.

TODO more than 15 sounds, non GM sounds, not just default instrument sound.