Music.Score.Harmonics

Description

Provides a representation of harmonics.

Harmonics

class HasHarmonic a where

Minimal complete definition

Nothing

Methods

setNatural :: Bool -> a -> a

default setNatural :: forall f b. (a ~ f b, Functor f, HasHarmonic b) => Bool -> a -> a

setHarmonic :: Int -> a -> a

default setHarmonic :: forall f b. (a ~ f b, Functor f, HasHarmonic b) => Int -> a -> a

data Harmonics

Constructors

NaturalHarmonic Natural

ArtificialHarmonic Natural

newtype HarmonicT a

Constructors

HarmonicT

Fields

getHarmonicT :: Couple (Any, Sum Int) a

runHarmonicT :: HarmonicT a -> (Harmonics, a)

harmonic :: HasHarmonic a => Int -> a -> a

Make all notes natural harmonics on the given overtone (1 for octave, 2 for fifth etc). Sounding pitch is unaffected, but notated output is transposed automatically.

artificial :: HasHarmonic a => a -> a

Make all notes natural harmonics on the given overtone (1 for octave, 2 for fifth etc). Sounding pitch is unaffected, but notated output is transposed automatically.