Music.Score.Meta.Annotations

Description

Provides a basic way annotating scores.

Documentation

data Annotation

An annotation is a unique textual value attached to parts of a score. They are ignored by default, but can be collected with withAnnotations.

getAnnotation :: Annotation -> [String]

annotate :: String -> Score a -> Score a

Annotate the whole score.

annotateSpan :: Span -> String -> Score a -> Score a

Annotate a part of the score.

showAnnotations :: (HasParts' a, Ord (Part a), HasText a) => Score a -> Score a

Show all annotations in the score.

showAnnotations' :: (HasParts' a, Ord (Part a), HasText a) => String -> Score a -> Score a

Show all annotations in the score using the given prefix.

withAnnotations :: ([String] -> Score a -> Score a) -> Score a -> Score a

Handle the annotations in a score.