Music.Score.Meta.Attribution

Description

Provides meta-data attribution for composer, lyricist etc.

Attribution type

data Attribution

An attributions is a simple map from keys to values used to gather information such as composer, lyricist, orchestrator, performer, etc.

Attribution is a Semigroup, and compose by choosing the leftmost value in each category. For example

attribution "composer" "H" <> attribution "composer" "S" <> attribution "lyricist" "S"
    ===> attributions [("composer","H"),("lyricist","S")]

Any kind of attribution can be added, and backends may recognize or ignore categories as they see fit. The following categories are normally recognized:

composer
lyricist
arranger
performer
dedication
year
copyright
information

attribution :: String -> String -> Attribution

Make an Attribution a single key and value.

attributions :: [(String, String)] -> Attribution

Make an Attribution from keys and values.

getAttribution :: Attribution -> String -> Maybe String

Extract an the given attributions value. Semantic function.

Adding attribution to scores

attribute :: (HasMeta a, HasPosition a) => Attribution -> a -> a

Set the given attribution in the given score.

attributeDuring :: HasMeta a => Span -> Attribution -> a -> a

Set the given attribution in the given part of a score.

composer :: (HasMeta a, HasPosition a) => String -> a -> a

Set composer of the given score.

composerDuring :: HasMeta a => Span -> String -> a -> a

Set composer of the given part of a score.

lyricist :: (HasMeta a, HasPosition a) => String -> a -> a

Set lyricist of the given score.

lyricistDuring :: HasMeta a => Span -> String -> a -> a

Set lyricist of the given part of a score.

arranger :: (HasMeta a, HasPosition a) => String -> a -> a

Set arranger of the given score.

arrangerDuring :: HasMeta a => Span -> String -> a -> a

Set arranger of the given part of a score.