Music.Score.Meta.Key

Description

Provides key signature meta-data.

Key signature type

newtype Fifths

Constructors

Fifths

Fields

getFifths :: Integer

newtype KeySignature

A key signature, represented by number of fifths from C and mode.

Constructors

KeySignature

Fields

getKeySignature :: First (Pitch, MajorMinor)

key :: Pitch -> MajorMinor -> KeySignature

Create a major or minor signature.

Adding key signatures to scores

keySignature :: (HasMeta a, HasPosition a) => KeySignature -> a -> a

Set the key signature of the given score.

keySignatureDuring :: HasMeta a => Span -> KeySignature -> a -> a

Set the key signature of the given part of a score.