Music.Score.Meta.Tempo

Description

Provides tempo meta-data.

Tempo type

type Bpm = Duration

type NoteValue = Duration

data Tempo

Represents musical tempo as a scaling factor with an optional name and/or beat duration.

tempoToDuration provides a scaling factor such that

stretch (tempoToDuration t) notation = sounding
compress (tempoToDuration t) sounding = notation

You can construct a tempo in various ways

tempoToDuration (metronome (1/4) 120) == tempoToDuration (metronome (1/2) 60)
metronome (1/4) 120                   /=                  metronome (1/2) 60

metronome :: Duration -> Bpm -> Tempo

Create a tempo from a duration and a number of beats per minute.

For example metronome (1/2) 48 means 48 half notes per minute.

tempoNoteValue :: Tempo -> Maybe NoteValue

Get the note value indicated by a tempo.

tempoBeatsPerMinute :: Tempo -> Bpm

Get the number of beats per minute indicated by a tempo.

getTempo :: Tempo -> (NoteValue, Bpm)

Get the note value and number of beats per minute indicated by a tempo.

Typically used with the ViewPatterns extension, as in

foo (getTempo -> (nv, bpm)) = ...

tempoToDuration :: Tempo -> Duration

Convert a tempo to a duration suitable for converting written to sounding durations.

stretch (tempoToDuration t) notation = sounding
compress (tempoToDuration t) sounding = notation

Adding tempo to scores

tempo :: (HasMeta a, HasPosition a) => Tempo -> a -> a

Set the tempo of the given score.

tempoDuring :: HasMeta a => Span -> Tempo -> a -> a

Set the tempo of the given part of a score.

Common tempi

presto :: Tempo

allegro :: Tempo

allegretto :: Tempo

moderato :: Tempo

andante :: Tempo

adagio :: Tempo

largo :: Tempo

lento :: Tempo