Music.Score.Meta.Time

Description

Provides time signatures and related meta-data.

Time signature type

newtype TimeSignature

A time signature is a sequence of beat numbers and a note value (i.e. an expression on the form (a1+a2...)/b). For simple time signatures just one beat number is used.

TimeSignature is an instance of Fractional and can be used as follows:

timeSignature (4/4)
timeSignature (6/8)
timeSignature ((3+2)/4)

Constructors

TimeSignature

Fields

getTimeSignature :: ([Integer], Integer)

time :: Integer -> Integer -> TimeSignature

Create a simple time signature.

compoundTime :: [Integer] -> Integer -> TimeSignature

Create a compound time signature.

isSimpleTime :: TimeSignature -> Bool

Whether this is a simple time signature.

isCompoundTime :: TimeSignature -> Bool

Whether this is a compound time signature.

toSimpleTime :: TimeSignature -> TimeSignature

Convert to a simple time signature by adding all numerators. If given a simple time signature, returns it.

Adding time signature to scores

timeSignature :: (HasMeta a, HasPosition a, Transformable a) => TimeSignature -> a -> a

Set the time signature of the given score.

timeSignatureDuring :: HasMeta a => Span -> TimeSignature -> a -> a

Set the time signature of the given part of a score.

Utility

standardTimeSignature :: Duration -> Maybe TimeSignature

Time signature typically used for the given duration.

Returns Nothing if the denominator of the canonical form of given duration is not a power of two.

TODO partial