Music.Score.Pitch

Description

Provides generic functions for inspecting and manipulating pitch.

The pitches will usually be of type Pitch, as defined in Music.Pitch.Common, but it is also possible to use other types such as Hertz.

Pitch functions

Transposition

up :: Transposable a => Interval a -> a -> a

Transpose (translate) up.

>>> up m3 (c :: Pitch)
eb

>>> up _P5 [c,d,e :: Pitch]
[g,a,b]

down :: Transposable a => Interval a -> a -> a

Transpose (translate) down.

>>> down m3 (c :: Pitch)
a

>>> down _P5 [c,d,e]
[f_,g_,a_]

above :: (Semigroup a, Transposable a) => Interval a -> a -> a

Add the given interval above.

>>> above _P5 [c :: Pitch]
[c,g]

>>> above _P5 (c :: Score Pitch)
[c,g]^.score

below :: (Semigroup a, Transposable a) => Interval a -> a -> a

Add the given interval below.

>>> below _P8 [c :: Pitch]
[c,c_]

octavesUp :: Transposable a => Scalar (Interval a) -> a -> a

Transpose up by the given number of octaves.

>>> octavesUp 2 (c :: Pitch)
c''

>>> octavesUp (-1) [c,d,e]
[c_,d_,e_]

octavesDown :: Transposable a => Scalar (Interval a) -> a -> a

Transpose down by the given number of octaves.

>>> octavesDown 2 (c :: Pitch)
c__

>>> octavesDown (-1) [c,d,e]
[c',d',e']

_15va :: Transposable a => a -> a

Same as octavesUp 2.

_8va :: Transposable a => a -> a

Same as octavesUp 1.

_8vb :: Transposable a => a -> a

Same as octavesDown 1.

_15vb :: Transposable a => a -> a

Same as octavesDown 2.

upDiatonic

Arguments

:: (HasPitches' a, Pitch a ~ Pitch)

=> Pitch
Tonic

-> DiatonicSteps
Number of steps to transpose

-> a
Original music

-> a

Diatonic transposition, using the diatonic scale centered around the given note.

>>> upDiatonic c 1 (e :: Pitch)
f
>>> upDiatonic g 1 (e :: Pitch)
fs
>>> upDiatonic c 2 [e,f,g :: Pitch]
[g,a,b]
>>> upDiatonic f 2 [e,f,g :: Pitch]
[g,a,bb]

downDiatonic

Arguments

:: (HasPitches' a, Pitch a ~ Pitch)

=> Pitch
Tonic

-> DiatonicSteps
Number of steps to transpose

-> a
Original music

-> a

Diatonic transposition, using the diatonic scale centered around the given note.

>>> upDiatonic c 1 (e :: Pitch)
f
>>> upDiatonic g 1 (e :: Pitch)
fs
>>> upDiatonic c 2 [e,f,g :: Pitch]
[g,a,b]
>>> upDiatonic f 2 [e,f,g :: Pitch]
[g,a,bb]

Inversion

invertPitches :: Transposable a => Pitch a -> a -> a

Invert pitches.

invertDiatonic

Arguments

:: (HasPitches' a, Pitch a ~ Pitch)

=> Pitch
Tonic

-> a
Original music

-> a

>>> invertDiatonic c ([e,gs]^.score :: Score Pitch)
[(0 <-> 1,a_)^.event,(0 <-> 1,fs_)^.event]^.score

invertChromatic

Arguments

:: (HasPitches' a, Pitch a ~ Pitch)

=> Pitch
Tonic

-> a
Original music

-> a

>>> invertChromatic c ([e,gs]^.score :: Score Pitch)
[(0 <-> 1,e)^.event,(0 <-> 1,gb)^.event]^.score

Ambitus

pitchRange :: (HasPitches' a, Ord p, p ~ Pitch a) => a -> Maybe (Ambitus v p)

highestPitch :: (HasPitches' a, Ord (Pitch a)) => a -> Maybe (Pitch a)

Extract the highest pitch. Returns Nothing if there are none.

>>> highestPitch [c,d,e :: Pitch]
Just e

>>> highestPitch (Data.Map.fromList [("do",c),("re",d)] :: Data.Map.Map String Pitch)
Just d

lowestPitch :: (HasPitches' a, Ord (Pitch a)) => a -> Maybe (Pitch a)

Extract the lowest pitch. Returns Nothing if there are none.

>>> highestPitch [c,d,e :: Pitch]
Just c

>>> highestPitch (Data.Map.fromList [("do",c),("re",d)] :: Data.Map.Map String Pitch)
Just c

averagePitch :: (HasPitches' a, Fractional (Pitch a)) => a -> Maybe (Pitch a)

Extract the average pitch. Returns Nothing if there are none.

>>> averagePitch (Data.Map.fromList [(True,440::Hertz),(False,445)])
Just 442.5 Hz

ambitusOctaves :: Ambitus Interval Pitch -> Int

The number of whole octaves in an ambitus.

ambitusLowestOctave :: Ambitus Interval Pitch -> Int

The lowest octave (relative middle C) in present a given ambitus.

interpolateAmbitus :: AffinePair v p => Ambitus v p -> Scalar v -> p

Interpolate between the highest and lowest points in an ambitus.

Can be used as a primitive contour-based melody generator.

interpolateAmbitus' :: Ambitus Double Pitch -> Double -> Pitch

Same as interpolateAmbitus but allow continous interpolation of standard pitch (as Scalar (Diff Pitch) ~ Integer).

Voices

stitch :: Transposable a => Voice a -> Voice a -> Voice a

Join two voices together so that one note overlaps. The second voice is transposed to achieve this.

At the join point the note from the first voice is used. If a ~ Pitch, then stitch and stitchLast are equivalent.

>>> stitch ([c :: Note Pitch,d,e]^.voice) ([c,g]^.voice)
[(1,c)^.note,(1,d)^.note,(1,e)^.note,(1,b)^.note]^.voice

stitchLast :: Transposable a => Voice a -> Voice a -> Voice a

stitchWith :: forall a. Transposable a => (Note a -> Note a -> Voice a) -> Voice a -> Voice a -> Voice a

Spelling

simplifyPitches :: (HasPitches' a, Pitch a ~ Pitch) => a -> a

Simpify the spelling of each pitch.

Enumeration

enumDiatonicFromTo :: Pitch -> Pitch -> [Pitch]

>>> enumDiatonicFromTo c c
[c]
>>> enumDiatonicFromTo f f'
[f,g,a,bb,c',d',e',f']
>>> 
>>> enumChromaticFromTo c c'
[c,cs,d,ds,e,f,fs,g,gs,a,as,b,c']
>>> 
>>> enumChromaticFromTo bs bs'
[bs,bss,css',csss',dss',es',ess',fss',fsss',gss',gsss',ass',bs']

enumChromaticFromTo :: Pitch -> Pitch -> [Pitch]

enumDownDiatonicFromTo :: Pitch -> Pitch -> [Pitch]

enumDownChromaticFromTo :: Pitch -> Pitch -> [Pitch]

Pitch type

type family Pitch (s :: Type) :: Type

A type function that returns the pitch type associated with a given type.

For simple types this is identity (they are their own pitch).

GetPitch Pitch = Pitch

For containers this is a morhism

GetPitch (Voice Pitch) = PitchOf Pitch = Pitch

type family SetPitch (b :: Type) (s :: Type) :: Type

A type function to change the pitch type associate with a given type, where the first argument is the new pitch type, and the second argument is the previous compound type.

For simple types this is constant (replacing means replacing the whole type).

SetPitch Hertz Pitch = Hertz

For containers this is a morhism

SetPitch Hertz (Voice Pitch) = Voice Hertz

type Interval a = Diff (Pitch a)

Associated interval type.

HasPitch classes

class HasPitches s t => HasPitch s t where

Types which has a single pitch (i.e notes, events, the pitches themselves).

Methods

pitch :: Lens s t (Pitch s) (Pitch t)

Access the pitch.

class SetPitch (Pitch t) s ~ t => HasPitches s t where

Types which has multiple pitches (i.e. voices, scores).

Methods

pitches :: Traversal s t (Pitch s) (Pitch t)

Access all pitches.

Simple versions

type HasPitch' a = HasPitch a a

Same as HasPitch but does not allow you to change the type.

type HasPitches' a = HasPitches a a

Same as HasPitches but does not allow you to change the type.

pitch' :: HasPitch' s => Lens' s (Pitch s)

Access the pitch. This is the same as pitch, but does not allow you to change the type.

pitches' :: HasPitches' s => Traversal' s (Pitch s)

Access all pitches. Same as pitches, but does not allow you to change the type.

Utility classes

type Transposable a = (HasPitches' a, AffinePair (Interval a) (Pitch a), PitchPair (Interval a) (Pitch a))

Class of types that can be transposed, inverted and so on.

type PitchPair v w = (Num (Scalar v), IsInterval v, IsPitch w)

type AffinePair v p = (VectorSpace v, AffineSpace p, Diff p ~ v)