Music.Score.Slide

Description

Provides a representation of musical slides, commonly known as glissando or portamento.

Slides and glissando

class HasSlide a where

Minimal complete definition

Nothing

Methods

setBeginGliss :: Bool -> a -> a

default setBeginGliss :: forall f b. (a ~ f b, Functor f, HasSlide b) => Bool -> a -> a

setBeginSlide :: Bool -> a -> a

default setBeginSlide :: forall f b. (a ~ f b, Functor f, HasSlide b) => Bool -> a -> a

setEndGliss :: Bool -> a -> a

default setEndGliss :: forall f b. (a ~ f b, Functor f, HasSlide b) => Bool -> a -> a

setEndSlide :: Bool -> a -> a

default setEndSlide :: forall f b. (a ~ f b, Functor f, HasSlide b) => Bool -> a -> a

newtype SlideT a

Constructors

SlideT

Fields

getSlideT :: Couple ((Any, Any), (Any, Any)) a

runSlideT :: SlideT a -> ((), a)

slide :: (HasPhrases' s a, HasSlide a) => s -> s

Add a slide between the first and the last note.

glissando :: (HasPhrases' s a, HasSlide a) => s -> s

Add a glissando between the first and the last note.