Music.Time.Aligned

Aligned type

data Aligned v

Aligned places a vector-like object in space, by fixing a local duration interpolating the vector to a specific point in time. The aligned value must be an instance of HasDuration, with view duration providing the size of the vector.

This is analogous to alignment in a graphical program. To align something at onset, midpoint or offset, use 0, 0.5 or 1 as the local duration value.

Creating aligned values

aligned :: Time -> Alignment -> v -> Aligned v

Align so that the given part of the value occurs at the given time.

aligned t 0 = onsetAt t
aligned t 1 = offsetAt t

alignTo :: (Transformable a, HasDuration a) => Time -> Duration -> a -> Aligned a

Align so that the given local duration occurs at the given time.

If _duration x = 1 then aligned t d x = alignTo t d x

(||>) :: (Transformable a, HasDuration a, Semigroup a) => a -> a -> Aligned a

Upbeat composition.

a ||> b is the same as a <> b aligned to the offset of b (and therefore also onset of b).

Update aligned values

align :: Alignment -> Aligned a -> Aligned a

Update the local origin of a value, but not its attachment point.

If the new alignment is different from the previous alignment, the value is moved.

align x (aligned t y) = aligned t x
align x (align y v) = align x v

realign :: (HasDuration a, Transformable a) => Alignment -> Aligned a -> Aligned a

Change the alignment of a value without moving it.

 x^.era = (realign l x)^.era
 

Extracting/rendering aligned values

renderAligned :: (HasDuration a, Transformable a) => (Span -> a -> b) -> Aligned a -> b

Render an aligned value. The given span represents the actual span of the aligned value.

renderAlignedVoice :: Aligned (Voice a) -> Score a

Convert an aligned voice to a score.

renderAlignedNote :: Aligned (Note a) -> Event a

Convert an aligned note to an event.

renderAlignedDuration :: Aligned Duration -> Span

Convert an aligned duration to a span.

unAlign :: Aligned a -> a

Conversion functions

voiceAtDuration :: Voice a -> Duration -> Maybe a

Treat a voice as a partial function of time.

>>> voiceAtDuration (stretch 2 c <> stretch 3 d) (-1)
Nothing

>>> voiceAtDuration (stretch 2 c <> stretch 3 d)) 0
Just c

>>> voiceAtDuration (stretch 2 c <> stretch 3 d) 2
Just d

>>> voiceAtDuration (stretch 2 c <> stretch 3 d) 6
Nothing

voiceToBehavior :: Monoid a => Aligned (Voice a) -> Behavior a

scoreToBehavior :: Monoid a => Score a -> Behavior a