Music.Time.Behavior

Behavior type

data Behavior a

A time-dependent value.

Construction

behavior :: Iso (Time -> a) (Time -> b) (Behavior a) (Behavior b)

View a behavior as a time function and vice versa.

Note that this is just an alias defined to make the documentation nicer:

behavior = tabulated

sampled :: Iso (Behavior a) (Behavior b) (Time -> a) (Time -> b)

View a time function as a behavior.

sampled    = from behavior
x^.sampled = (x !)

Lookup

(!) :: Behavior a -> Time -> a

Sample the given behavior at the given time.

Combinators

switch :: Time -> Behavior a -> Behavior a -> Behavior a

Instantly switch from one behavior to another.

switch' :: Time -> Behavior a -> Behavior a -> Behavior a -> Behavior a

Instantly switch from one behavior to another with an optinal intermediate value.

trim :: Monoid a => Span -> Behavior a -> Behavior a

Replace everything outside the given span by mempty.

trimBefore :: Monoid a => Time -> Behavior a -> Behavior a

Replace everthing before the given time by mempty.

trimAfter :: Monoid a => Time -> Behavior a -> Behavior a

Replace everthing after the given time by mempty.

concatB :: Monoid a => Score (Behavior a) -> Behavior a

Flatten a score of behaviors. The behavior for each note is sampled in the units span 0 - 1. Overlapping notes are merged using the monoid instance.

concatB is a monoid homomorphism, that is:

concatB mempty   = mempty
concatB (a <> b) = concatB a <> concatB b

Common behaviors

Oscillators

line :: Fractional a => Behavior a

A behavior that gives the current time.

(line !) = id

sawtooth :: RealFrac a => Behavior a

A behavior that goes from 0 to 1 repeatedly with a period of 1.

sine :: Floating a => Behavior a

A behavior that

cosine :: Floating a => Behavior a

A behavior that

Impulse functions

unit :: Fractional a => Behavior a

A behavior that varies from 0 to 1 during the same time interval and is 0 before and 1 after that interval.

impulse :: Num a => Behavior a

A behavior that is 1 at time 0, and 0 at all other times.

turnOn :: Num a => Behavior a

A behavior that goes from 0 to 1 at time 0.

turnOff :: Num a => Behavior a

A behavior that goes from 1 to 0 at time 0.