Music.Time.Event

Event type

data Event a

A Event is a value transformed to appear in some Span. Like Span, it is an instance of Transformable.

Construction

event :: Iso (Span, a) (Span, b) (Event a) (Event b)

View a event as a pair of the original value and the transformation (and vice versa).

eventee :: (Transformable a, Transformable b) => Lens (Event a) (Event b) a b

View the value in the event.

spanEvent :: Iso' Span (Event ())

Event as a span with a trivial value.