Music.Time.Juxtapose

Documentation

module Music.Time.Split

Align without composition

lead :: (HasPosition a, HasPosition b, Transformable a) => a -> b -> a

(a lead b)^.offset = b^.onset

follow :: (HasPosition a, HasPosition b, Transformable b) => a -> b -> b

a^.offset = (a follow b)^.onset

Standard composition

after :: (Semigroup a, Transformable a, HasPosition a) => a -> a -> a

before :: (Semigroup a, Transformable a, HasPosition a) => a -> a -> a

during :: (HasPosition a, HasPosition b, Transformable a) => a -> b -> a

Move a value so that its era is equal to the era of another value.

(|>) :: (Semigroup a, HasPosition a, Transformable a) => a -> a -> a infixr 6

An infix alias for after.

(<|) :: (Semigroup a, HasPosition a, Transformable a) => a -> a -> a infixr 6

An infix alias for before.

More exotic

rev :: (Transformable a, HasPosition a) => a -> a

sustain :: (Semigroup a, HasPosition a, Transformable a) => a -> a -> a

Like <>, but scaling the second agument to the duration of the first.

palindrome :: (Semigroup a, Transformable a, HasPosition a) => a -> a

A value followed by its reverse (retrograde).

Catenation

pseq :: (Monoid a, HasPosition a, Transformable a) => [a] -> a

Compose a list of sequential objects, with onset and offset tangent to one another.

ppar :: Monoid a => [a] -> a

Compose a list of parallel objects, so that their local origins align.

For positioned types this is the same as mconcat.

Repetition

times :: (Monoid a, HasPosition a, Transformable a) => Int -> a -> a

Repeat the given music n times.

Int -> Score a -> Score a

group :: (Monoid a, Transformable a, HasPosition a) => Int -> a -> a

Repeat the given music n times and stretch to the original duration.

Int -> Score a -> Score a

newtype After a

Monoid under sequential composition.

Warning: This is only lawful if the underlying monoid means "parallel composition". E.g. this works for Score and Pattern, but not for Span and Voice.

Constructors

After

Fields

getAfter :: a