Music.Time.Meta

Description

Provides a way to annotate data-types with Transformable meta-data. Inspired by Clojure meta-data and Diagrams styles.

Attributes

type AttributeClass a = (Typeable a, Monoid a, Semigroup a)

Class of values that can be wrapped.

type TAttributeClass a = (Transformable a, AttributeClass a)

Class of values that can be wrapped and transformed.

data Attribute

An existential wrapper type to hold attributes.

Creating attributes

wrapAttr :: AttributeClass a => a -> Attribute

Wrap up an attribute.

wrapTAttr :: TAttributeClass a => a -> Attribute

Wrap up a transformable attribute.

unwrapAttr :: AttributeClass a => Attribute -> Maybe a

Convert something from an attribute. Also works with transformable attributes

Meta-data

data Meta

Creating meta-data

wrapMeta :: forall a. AttributeClass a => a -> Meta

Convert something from meta-data. Also works with transformable attributes

wrapTMeta :: forall a. TAttributeClass a => a -> Meta

Convert something to meta-data.

unwrapMeta :: forall a. AttributeClass a => Meta -> Maybe a

Convert something from meta-data.

The HasMeta class

class HasMeta a where

Type class for things which have meta-data.

Laws: meta is a lens, e.g. view meta (set meta v s) ≡ v set meta (view meta s) s ≡ s set meta v' (set meta v s) ≡ set meta v' s

Methods

meta :: Lens' a Meta

Access the meta-data.

getMeta :: HasMeta a => a -> Meta

Extract meta-data.

mapMeta :: HasMeta a => (Meta -> Meta) -> a -> a

Map over meta-data.

setMeta :: HasMeta a => Meta -> a -> a

Update meta-data.

metaTypes :: HasMeta a => a -> [String]

Show the types of meta-data attachd to this value. Useful for debugging.

applyMeta :: HasMeta a => Meta -> a -> a

Apply meta-information by combining it with existing meta-information.

setMetaAttr :: (AttributeClass b, HasMeta a) => b -> a -> a

Update a meta attribute.

setMetaTAttr :: (TAttributeClass b, HasMeta a) => b -> a -> a

Update a meta attribute.

preserveMeta :: (HasMeta a, HasMeta b) => (a -> b) -> a -> b

Apply a function without affecting meta-data.

preserveMetaF :: (HasMeta a, HasMeta b, Functor f) => (a -> f b) -> a -> f b

Apply a function without affecting meta-data.

Add meta-data to arbitrary types

data AddMeta a

Annotate an arbitrary type with meta-data, preserving instances of all common type classes. In particular Functor and Applicative is lifted and Compose AddMeta is semantically equivalent to Identity.

annotated :: Lens (AddMeta a) (AddMeta b) a b

Access the annotated value.

over annotated = fmap

unannotated :: Getter a (AddMeta a)

Access the annotated value.

view fromAnnotated = pure

annotatedIgnoringMeta :: Iso (AddMeta a) (AddMeta b) a b

Access the annotated value. This is only an isomorphism up to meta-data equivalence. In particular under annotatedIgnoringMeta leads to meta-data being thrown away. See annotated and unannotated for safe (but less general) definitions.

over annotated = fmap