Music.Time.Pattern

Documentation

data Pattern a

newPattern :: Voice a -> Pattern a

rhythmPattern :: IsPitch a => [Duration] -> Pattern a

renderPattern :: Pattern a -> Span -> Score a

renderPatternsRel :: Score (Pattern a) -> Score a

Renders each pattern over one cycle (0-1) and stretches to the duration of the surrounding note.

Each note triggers exactly one cycle of the pattern (frequency = 1), starting at the beginning of the pattern (phase = 0).

renderPatternsAbs :: Score (Pattern a) -> Score a

Renders each pattern in the span of its note.

This means that notes of different onset and duration may trigger a different number of cycles (frequency), with different starting point in the pattern (phase).