Music.Time.Position

Documentation

module Music.Time.Duration

The HasPosition class

class HasPosition a where

Class of values that have a position in time.

Laws

duration-position

For HasDuration instances:

fmap _duration . _era = Just . _duration

transform-position

For Transformable instances:

_era (transform s x) = fmap (transform s) (_era x)

Methods

_era :: a -> Maybe Span

Return the conventional bounds of a value (local time zero and one).

class HasPosition a => HasPosition1 a where

Class of values that have a position in time.

Laws

duration-position

For HasDuration instances:

_duration . _era1 = _duration

transform-position

For Transformable instances:

_era1 (transform s x) = transform s (_era1 x)

Methods

_era1 :: a -> Span

_position :: HasPosition a => a -> Duration -> Maybe Time

Map a local time in value to global time.

_position1 :: HasPosition1 a => a -> Duration -> Time

Map a local time in value to global time.

Position and Era

position :: (HasPosition1 a, Transformable a) => Duration -> Lens' a Time

Position of the given value.

era :: (HasPosition1 a, Transformable a) => Lens' a Span

setEra :: (HasPosition a, Transformable a) => Span -> a -> a

Place a value over the given span.

placeAt s t places the given thing so that x^.place = s

Specific positions

onset :: (HasPosition1 a, Transformable a) => Lens' a Time

Onset of the given value, corresponding to alignment 0.

midpoint :: (HasPosition1 a, Transformable a) => Lens' a Time

Midpoint of the given value, or the value between the decay and sustain phases.

offset :: (HasPosition1 a, Transformable a) => Lens' a Time

Offset of the given value, corresponding to alignment 1.

preOnset :: (HasPosition1 a, Transformable a) => Lens' a Time

Pre-onset of the given value, or the value right before the attack phase.

postOffset :: (HasPosition1 a, Transformable a) => Lens' a Time

Post-offset of the given value, or the value right after the release phase.

Moving

startAt :: (Transformable a, HasPosition a) => Time -> a -> a

Move a value forward in time.

stopAt :: (Transformable a, HasPosition a) => Time -> a -> a

Move a value forward in time.

placeAt :: (Transformable a, HasPosition a) => Alignment -> Time -> a -> a

Align a value to a given position.

placeAt p t places the given thing so that its position p is at time t

placeAt 0 = startAt
placeAt 1 = stopAt

Transforming

stretchTo :: (Transformable a, HasPosition a) => Duration -> a -> a

Stretch a value to have the given duration.

Transforming relative a position

stretchRelative :: (HasPosition a, Transformable a) => Alignment -> Duration -> a -> a

Stretch a value relative to its local origin.

stretchRelativeOnset    = stretchRelative 0
stretchRelativeMidpoint = stretchRelative 0.5
stretchRelativeOffset   = stretchRelative 1

stretchRelativeOnset :: (HasPosition a, Transformable a) => Duration -> a -> a

Stretch a value relative to its onset.

>>> stretchRelativeOnset 2 (0 <-> 1)
0 <-> 2

stretchRelativeMidpoint :: (HasPosition a, Transformable a) => Duration -> a -> a

Stretch a value relative to its midpoint.

>>> stretchRelativeMidpoint 2 (1 <-> 3)
0 <-> 4

stretchRelativeOffset :: (HasPosition a, Transformable a) => Duration -> a -> a

Stretch a value relative to its offset.

>>> stretchRelativeOffset 2 (1 <-> 2)
0 <-> 2

transformRelative :: (HasPosition a, Transformable a) => Alignment -> Span -> a -> a

Transform a value relative to its local origin.

stretchRelativeOnset    = stretchRelative 0
stretchRelativeMidpoint = stretchRelative 0.5
stretchRelativeOffset   = stretchRelative 1

transformRelativeOnset :: (HasPosition a, Transformable a) => Span -> a -> a

Transform a value relative to its onset.

transformRelativeMidpoint :: (HasPosition a, Transformable a) => Span -> a -> a

Transform a value relative to its midpoint.

transformRelativeOffset :: (HasPosition a, Transformable a) => Span -> a -> a

Transform a value relative to its offset.