Music.Time.Score

Score type

data Score a

A set of events of type a, with an associated time spans.

Construction

events . each . from event . swapped . mapping onsetAndDuration . swapped
   :: Traversal (Score a) (Score b) ((Time, Duration), a) ((Time, Duration), b)

events . each . from event . swapped . mapping onsetAndOffset . swapped
   :: Traversal (Score a) (Score b) ((Time, Time), a) ((Time, Time), b)

score :: Getter [Event a] (Score a)

Create a score from a list of events.

events :: Lens (Score a) (Score b) [Event a] [Event b]

View a Score as a list of Event values.

eras :: Traversal' (Score a) Span

Print all eras of the given score.

>>> toListOf eras $ pseq [c,d,e :: Score Integer]
[0 <-> 1,1 <-> 2,2 <-> 3]

triples :: Lens (Score a) (Score b) [(Time, Duration, a)] [(Time, Duration, b)]

View a score as a list of time-duration-value triplets.

TODO replace this with traversals

Traversal

mapWithSpan :: (Span -> a -> b) -> Score a -> Score b

Map over the values in a score.

filterWithSpan :: (Span -> a -> Bool) -> Score a -> Score a

Filter the values in a score.

mapFilterWithSpan :: (Span -> a -> Maybe b) -> Score a -> Score b

Efficient combination of mapWithSpan and filterWithTime.

mapWithTime :: (Time -> Duration -> a -> b) -> Score a -> Score b

Map over the values in a score.

filterWithTime :: (Time -> Duration -> a -> Bool) -> Score a -> Score a

Filter the values in a score.

mapFilterWithTime :: (Time -> Duration -> a -> Maybe b) -> Score a -> Score b

Efficient combination of mapWithTime and filterWithTime.

Simultaneous

hasOverlappingEvents :: Score a -> Bool

True if the score has more than one overlapping event.

hasOverlappingEvents (simultaneous x) = False

simultaneous :: (Transformable a, Semigroup a) => Score a -> Score a

Merge all simultaneous events using their Semigroup instance.

Normalize

normalizeScore :: Score a -> Score a

Normalize a score, assuring its events spans are all forward (as by isForwardSpan), and that its onset is at least zero. Consequently, the onset and offset of each event in the score is at least zero.

removeRests :: Score (Maybe a) -> Score a

Remove all Nothing values in the score.

Unsafe versions

eventsIgnoringMeta :: Iso (Score a) (Score b) [Event a] [Event b]

A score is a list of events up to meta-data. To preserve meta-data, use the more restricted score and events.

triplesIgnoringMeta :: Iso (Score a) (Score b) [(Time, Duration, a)] [(Time, Duration, b)]

A score is a list of (time-duration-value triples) up to meta-data. To preserve meta-data, use the more restricted triples.

Conversion

eventToScore :: Event a -> Score a

Convert an event to a singleton score.