Music.Time.Transform

Documentation

module Music.Time.Types

The Transformable class

class Transformable a where

Class of values that can be transformed (i.e. scaled and moved) in time.

Laws

identity

transform mempty = id

compatibility

transform (s <> t) = transform s . transform t

In other words, transform must be a left-action of Span on a.

Methods

transform :: Span -> a -> a

transformed :: (Transformable a, Transformable b) => Span -> Iso a b a b

View the given value in the context of the given transformation.

Specific transformations

Transformations

delaying :: Duration -> Span

A transformation that moves a value forward in time.

undelaying :: Duration -> Span

A transformation that moves a value backward in time.

stretching :: Duration -> Span

A transformation that stretches (augments) a value by the given factor.

compressing :: Duration -> Span

A transformation that compresses (diminishes) a value by the given factor.

Transforming values

delay :: Transformable a => Duration -> a -> a

Moves a value forward in time.

undelay :: Transformable a => Duration -> a -> a

Moves a value backward in time. Equnitvalent to stretch . negate.

stretch :: Transformable a => Duration -> a -> a

Stretches (augments) a value by the given factor.

compress :: Transformable a => Duration -> a -> a

Compresses (diminishes) a score. Equnitvalent to stretch . recip.

(|*) :: forall a. Transformable a => a -> Duration -> a infixl 7

Infix version of stretch.

(*|) :: forall a. Transformable a => Duration -> a -> a infixr 7

Infix version of stretch.

(|/) :: forall a. Transformable a => a -> Duration -> a infixr 7

Infix version of compress.