Music.Time.Types

Basic types

data Time

Time represents points in time space. The difference between two time points is a Duration, for example in a bar of duration 4/4 (that is 1), the difference between the first and third beat 1/2.

Time has an origin (zero) which usually represents the beginning of the musical performance, but this may not always be the case, as the modelled music may be infinite, or contain a musical pickup. Hence Time values can be negative.

Duration is isomorphic to Rational. You can use toRational and fromRational to convert it. To convert between Time and Duration, use the AffineSpace instance.

data Duration

Duration, corresponding to note values in standard notation. The standard names can be used: 1/2 for half note 1/4 for a quarter note and so on.

Duration is isomorphic to Rational. You can use toRational and fromRational to convert it. To convert between Time and Duration, use the AffineSpace instance.

type Alignment = Duration

Alignment is a synonym for Duration.

See Music.Time.Aligned for its intended use.

Convert between time and duration

Note that you should use .-. and .+^ to convert between time and duration. To refer to time zero (the beginning of the music), use origin.

toAbsoluteTime :: [Duration] -> [Time]

Interpret as durations from 0.

toAbsoluteTime (toRelativeTime xs) == xs

lenght xs == length (toRelativeTime xs)

>>> toAbsoluteTime [1,1,1] :: [Time]
[1,2,3]

toRelativeTime :: [Time] -> [Duration]

Duration between 0 and first value and so on until the last.

toAbsoluteTime (toRelativeTime xs) == xs

lenght xs == length (toRelativeTime xs)

>>> toRelativeTime [1,2,3]
[1,1,1]

toRelativeTimeN :: [Time] -> [Duration]

toRelativeTimeN' :: Time -> [Time] -> [Duration]

Duration between values until the last, then up to the given final value. > lenght xs == length (toRelativeTime xs)

Time spans

data Span

A Span represents a time interval. It has a starting point called onset and an ending point called offset.

Another way of looking at Span is that it represents a time transformation where onset is translation and duration is scaling.

This type is sometimes known as arc or era.

Constructing spans

(<->) :: Time -> Time -> Span infixl 6

t <-> u represents the span between t and u.

(>->) :: Time -> Duration -> Span infixl 6

t >-> d represents the span between t and t .+^ d.

(<-<) :: Duration -> Time -> Span infixl 6

d <-> t represents the span between t .-^ d and t.

onsetAndOffset :: Iso' Span (Time, Time)

View a span as pair of onset and offset.

onsetAndDuration :: Iso' Span (Time, Duration)

View a span as a pair of onset and duration.

durationAndOffset :: Iso' Span (Duration, Time)

View a span as a pair of duration and offset.

stretchComponent :: Span -> Duration

Access the stretch component in a span.

delayComponent :: Span -> Time

Access the delay component in a span.

fixedDurationSpan :: Prism' Span Time

A prism to the subset of Span that performs a delay but no stretch.

fixedOnsetSpan :: Prism' Span Duration

A prism to the subset of Span that performs a stretch but no delay.

Transformations

normalizeSpan :: Span -> Span

Normalize a span, i.e. reverse it if negative, and do nothing otherwise.

abs $ s^.duration = abs $ (normalizeSpan s)^.duration
s^.midpoint = (normalizeSpan s)^.midpoint

reverseSpan :: Span -> Span

Reflect a span through its midpoint.

reflectSpan :: Time -> Span -> Span

Reflect a span through an arbitrary point.

Properties

isDegenerateSpan :: Span -> Bool

Whether the given span is empty, i.e. whether its onset and offset are equivalent.

isForwardSpan :: Span -> Bool

Whether the given span has a positive duration, i.e. whether its onset is before its offset.

isBackwardSpan :: Span -> Bool

Whether the given span has a negative duration, i.e. whether its offset is before its onset.

Combine

hull :: Span -> Span -> Span

Return the convex hull of two spans.

This is associative. Its identity would be the empty span, but this Span represents non-empty time spans only.

data TimeInterval

Constructors

EmptyInterval

NonEmptyInterval Span

Points in spans

inside :: Time -> Span -> Bool infixl 5

Whether the given point falls inside the given span.

Designed to be used infix, for example

>>> 0.5 `inside` 1 <-> 2
False

>>> 1.5 `inside` 1 <-> 2
True

>>> 1 `inside` 1 <-> 2
True

>>> 2 `inside` 1 <-> 2
True

strictlyInside :: Time -> Span -> Bool

Whether the given point falls inside the given span.

>>> 2 `inside` (3<->4)
False
>>> 3 `inside` (3<->4)
True

>>> 2 `strictlyInside` (3<->4)
False
>>> 3 `strictlyInside` (3<->4)
False
>>> 3.5 `strictlyInside` (3<->4)
True

closestPointInside :: Span -> Time -> Time

If the given time is outside the given span, return the closest point inside. Otherwise return the given time.

>>> closestPointInside (1 <-> 3) 0
1
>>> closestPointInside (1 <-> 3) 55
3
>>> closestPointInside (1 <-> 3) 2
2

Predicates

encloses :: Span -> Span -> Bool infixl 5

Whether the first given span encloses the second span.

>>> 0 <-> 3 `encloses` 1 <-> 2
True

>>> 0 <-> 2 `encloses` 1 <-> 2
True

>>> 1 <-> 3 `encloses` 1 <-> 2
True

>>> 1 <-> 2 `encloses` 1 <-> 2
True

properlyEncloses :: Span -> Span -> Bool infixl 5

Whether the first given span encloses the second span.

>>> 0 <-> 3 `properlyEncloses` 1 <-> 2
True

>>> 0 <-> 2 `properlyEncloses` 1 <-> 2
True

>>> 1 <-> 3 `properlyEncloses` 1 <-> 2
True

>>> 1 <-> 2 `properlyEncloses` 1 <-> 2
False

overlaps :: Span -> Span -> Bool infixl 5

Whether the given span overlaps.

isBefore :: Span -> Span -> Bool

Whether the first given span occurs before the second span.

Read/Show

showOnsetAndOffset :: Span -> String

Show a span as a pair of onset and offset, i.e. t1 <-> t2.

showOnsetAndDuration :: Span -> String

Show a span as a pair of onset and duration, i.e. t >-> d.

showDurationAndOffset :: Span -> String

Show a span as a pair of duration and offset, i.e. t <-< d.