Music.Time.Voice

Voice type

data Voice a

A sequential composition of values with associated durations.

Construction

view notes                        :: Voice a -> [Note a]
set  notes                        :: [Note a] -> Voice a -> Voice a
over notes                        :: ([Note a] -> [Note b]) -> Voice a -> Voice b

preview  (notes . each)           :: Voice a -> Maybe (Note a)
preview  (notes . element 1)      :: Voice a -> Maybe (Note a)
preview  (notes . elements odd)   :: Voice a -> Maybe (Note a)

set      (notes . each)           :: Note a -> Voice a -> Voice a
set      (notes . element 1)      :: Note a -> Voice a -> Voice a
set      (notes . elements odd)   :: Note a -> Voice a -> Voice a

over     (notes . each)           :: (Note a -> Note b) -> Voice a -> Voice b
over     (notes . element 1)      :: (Note a -> Note a) -> Voice a -> Voice a
over     (notes . elements odd)   :: (Note a -> Note a) -> Voice a -> Voice a

toListOf (notes . each)                :: Voice a -> [Note a]
toListOf (notes . elements odd)        :: Voice a -> [Note a]
toListOf (notes . each . filtered
             (\x -> x^.duration < 2))  :: Voice a -> [Note a]

voice :: Iso' [Note a] (Voice a)

Create a Voice from a list of Notes.

notes :: Iso (Voice a) (Voice b) [Note a] [Note b]

View a Voice as a list of Note values.

pairs :: Iso (Voice a) (Voice b) [(Duration, a)] [(Duration, b)]

View a score as a list of duration-value pairs. Analogous to triples.

durationsAsVoice :: Iso' [Duration] (Voice ())

Map

mapWithOnsetRelative :: Time -> (Time -> a -> b) -> Voice a -> Voice b

mapWithOffsetRelative :: Time -> (Time -> a -> b) -> Voice a -> Voice b

mapWithSpanRelative :: Time -> (Span -> a -> b) -> Voice a -> Voice b

Points in a voice

onsetsRelative :: Time -> Voice a -> [Time]

Returns the onsets of all notes in a voice given the onset of the first note.

offsetsRelative :: Time -> Voice a -> [Time]

Returns the offsets of all notes in a voice given the onset of the first note.

midpointsRelative :: Time -> Voice a -> [Time]

Returns the midpoints of all notes in a voice given the onset of the first note.

erasRelative :: Time -> Voice a -> [Span]

Returns the eras of all notes in a voice given the onset of the first note.

Transformations

Rotation

rotateDurations :: Int -> Voice a -> Voice a

Rotate the durations of a voice.

>>> rotateDurations 1 [(1,'c'), (2, 'd'), (1, 'e')]
[(2,'c'), (1, 'd'), (1, 'e')]

rotateValues :: Int -> Voice a -> Voice a

Rotate the values of a voice.

>>> rotateValues 1 [(1,'c'), (2, 'd'), (1, 'e')]
[(1,'d'), (2, 'e'), (1, 'c')]

Fusion

fuse :: Eq a => Voice a -> Voice a

Merge consecutive equal notes.

fuseBy :: (a -> a -> Bool) -> Voice a -> Voice a

Merge consecutive notes deemed equal by the given predicate.

Fuse rests

fuseRests :: Voice (Maybe a) -> Voice (Maybe a)

Fuse all rests in the given voice. The resulting voice will have no consecutive rests.

coverRests :: Voice (Maybe a) -> Maybe (Voice a)

Remove all rests in the given voice by prolonging the previous note. Returns Nothing if and only if the given voice contains rests only.

Zips and unzip

zipVoiceScale :: Voice a -> Voice b -> Voice (a, b)

Join the given voices by multiplying durations and pairing values.

zipVoiceNoScale :: Voice a -> Voice b -> Voice (a, b)

Join the given voices by pairing values and selecting the first duration.

zipVoiceScaleWith :: (a -> b -> c) -> Voice a -> Voice b -> Voice c

Join the given voices by multiplying durations and combining values using the given function.

zipVoiceWithNoScale :: (a -> b -> c) -> Voice a -> Voice b -> Voice c

Join the given voices without combining durations. The durations are taken from the first voice.

zipVoiceWith' :: (Duration -> Duration -> Duration) -> (a -> b -> c) -> Voice a -> Voice b -> Voice c

Join the given voices by combining durations and values using the given function.

unzipVoice :: Voice (a, b) -> (Voice a, Voice b)

Unzip the given voice.

Merging

sameDurations :: Voice a -> Voice b -> Bool

Whether two notes have exactly the same duration pattern. Two empty voices are considered to have the same duration pattern. Voices with an non-equal number of notes differ by default.

mergeIfSameDuration :: Voice a -> Voice b -> Maybe (Voice (a, b))

Pair the values of two voices if and only if they have the same duration pattern (as per sameDurations).

mergeIfSameDurationWith :: (a -> b -> c) -> Voice a -> Voice b -> Maybe (Voice c)

Combine the values of two voices using the given function if and only if they have the same duration pattern (as per sameDurations).

homoToPolyphonic :: Voice [a] -> [Voice a]

Split a homophonic texture into a polyphonic one. The returned voice list will have as many elements as the chord with the fewest number of notes.

Context

withContext :: Voice a -> Voice (Ctxt a)

Decorate all notes in a voice with their context, i.e. previous and following value if present.

Conversion

noteToVoice :: Note a -> Voice a

Convert a note to a singleton voice.